Facing direction of player sent over server

+++ Arcade Arena/Character.cs
@@ -95,7 +95,7 @@ namespace Arcade_Arena
 
         public sbyte Health { get { return health; } private set { health = value; } }
 
-        public double OrbiterRotation { get { return orbiterRotation; } private set { orbiterRotation = value; } }
+        public float OrbiterRotation { get { return orbiterRotation; } private set { orbiterRotation = value; } }
 
         public bool IsDead => isDead;
